Fix sizing of Maps
Fix several instances where Maps were sized incorrectly when the number of
elements to insert is known; specifying the capacity without taking into
account the load factor meant it would always result in a rehash / resize
operation.

To this end, add a simple Maps utility class that handles the capacity
calculation, much like the static HashMap.newHashMap helper that was added in JDK19.