Raise default async discard threshold to ERROR
There are two basic scenarios where the ring buffer fills up. One is
that an application is simply logging too much and the flushing process
can't keep up, and in this case a discard threshold of WARN or INFO is
probably sufficient to mitigate the problem. However, if flushing has
stopped making progress altogether, e.g. due to a full or failed disk,
then logging calls will block indefinitely. This can result in a
production outage.

This change sets the default discard threshold to ERROR, in order to
better mitigate the scenario where the disk fills up, fails, or is in
the process of failing. With this threshold, logging should only block
at the FATAL level, which would typically mean that the operation is
already failing anyway.