Fix: prevent ThrowableStackTraceRenderer from throwing NPEs
The `ThrowableStackTraceRenderer` class can throw a `NullPointerException` if the suppressed
exceptions associated with the `Throwable` it is rendering are being concurrently mutated.
This happens because `ThrowableStackTraceRenderer` invokes `Throwable#getSuppressed()` twice:
once in `ThrowableStackTraceRenderer.Context.Metadata#populateMetadata()`, and a second time in
`ThrowableStackTraceRenderer#renderThrowable()`, ahead of invoking
`ThrowableStackTraceRenderer#renderSuppressed()`. If a racing thread manages to add a new
suppressed exception to the being-logged exception between these two invocations, then the
`Map<Throwable, Context.Metadata>` constructed by `populateMetadata()` will contain no mapping
for the newly-added suppression, and as a result the dereference performed on line 168 explodes.

Note: the unit test I am adding here requires the ability to mock `Throwable#getSuppressed()`.
Since this method is `final`, I had to add a dependency on `mockito-inline`, which then required
that I fix up some unrelated unit tests that had been relying on the lack of support for mocking
`final` methods.

This fixes #3929